chore: initial typescript migration with unit and integration tests
