+++ nodemon-debug.json
@@ -1,0 +1,8 @@
+{
+  "watch": [
+    "."
+  ],
+  "ext": "ts",
+  "ignore": ["**/*.spec.ts"],
+  "exec": "node --inspect=0.0.0.0 --require ts-node/register --require tsconfig-paths/register tonicExample.ts"
+}
