+++ test/mocha.opts
@@ -1,0 +1,6 @@
+--require ts-node/register
+--require source-map-support/register
+--recursive
+--sort
+--full-trace
+--exit
