--- tonicExample.js
@@ -1,3 +1,0 @@
-var bimserverapi = require("bimserverapi");
-
-bimserverapi.login("admin@bimserver.org", "admin");
