--- translations_en.js
@@ -1,74 +1,0 @@
-export const translations = {
-	GETDATAOBJECTSBYTYPE_BUSY: "Loading objects",
-	REQUESTPASSWORDCHANGE_BUSY: "Busy sending password reset e-mail",
-	REQUESTPASSWORDCHANGE_DONE: "A password reset e-mail has been sent",
-	SETSERVERSETTINGS_DONE: "Server settings successfully updated",
-	ENABLEPLUGIN_DONE: "Plugin successfully enabled",
-	DISABLEPLUGIN_DONE: "Plugin successfully disabled",
-	SETDEFAULTWEBMODULE_DONE: "Default webmodule changed",
-	SETDEFAULTQUERYENGINE_DONE: "Default Query Engine successfully changed",
-	SETDEFAULTMODELMERGER_DONE: "Default Model Merger successfully changed",
-	SETDEFAULTSERIALIZER_DONE: "Default Serializer successfully changed",
-	SETDEFAULTOBJECTIDM_DONE: "Default ObjectIDM successfully changed",
-	SETDEFAULTRENDERENGINE_DONE: "Default Render Engine successfully changed",
-	SETDEFAULTMODELCOMPARE_DONE: "Default Model Compare successfully changed",
-	LOGIN_BUSY: "Trying to login",
-	CHANGEUSERTYPE_DONE: "Type of user successfully changed",
-	ADDUSER_DONE: "User successfully added, you should receive a validation email shortly",
-	UPDATEINTERNALSERVICE_DONE: "Internal service successfully updated",
-	UPDATEMODELCOMPARE_DONE: "Model compare plugin successfully updated",
-	UPDATEMODELMERGER_DONE: "Model merger successfully updated",
-	UPDATEQUERYENGINE_DONE: "Query engine plugin successfully updated",
-	UPDATEOBJECTIDM_DONE: "ObjectIDM succesfully updated",
-	UPDATEDESERIALIZER_DONE: "Serializer succesfully updated",
-	ADDUSERTOPROJECT_DONE: "User successfully added to project",
-	REMOVEUSERFROMPROJECT_DONE: "User successfully removed from project",
-	UNDELETEPROJECT_DONE: "Project successfully undeleted",
-	DELETEPROJECT_DONE: "Project successfully deleted",
-	ADDPROJECT_DONE: "Project successfully added",
-	VALIDATEACCOUNT_DONE: "Account successfully validated, you can now login",
-	ADDPROJECTASSUBPROJECT_DONE: "Sub project added successfully",
-	DOWNLOADBYJSONQUERY_BUSY: "Downloading BIM",
-	CHECKINFROMURL_DONE: "Done checking in from URL",
-	GETLOGGEDINUSER_BUSY: "Getting user details",
-	SETPLUGINSETTINGS_DONE: "Plugin settings successfully saved",
-	GETSERVERINFO_BUSY: "Getting server info",
-	GETVERSION_BUSY: "Getting server version",
-	GETPROJECTBYPOID_BUSY: "Getting project details",
-	GETALLRELATEDPROJECTS_BUSY: "Getting related project's details",
-	GETSERIALIZERBYPLUGINCLASSNAME_BUSY: "Getting serializer info",
-	CLEANUPLONGACTION_BUSY: "Cleaning up",
-	GETREVISIONSUMMARY_BUSY: "Getting revision summary",
-	DOWNLOADBYOIDS_BUSY: "Downloading model data",
-	REGISTERPROGRESSHANDLER_BUSY: "Registering for updates on progress",
-	GETALLREVISIONSOFPROJECT_BUSY: "Getting all revisions of project",
-	GETPLUGINDESCRIPTOR_BUSY: "Getting plugin information",
-	GETUSERSETTINGS_BUSY: "Getting user settings",
-	GETALLQUERYENGINES_BUSY: "Getting query engines",
-	REGISTERNEWPROJECTHANDLER_BUSY: "Registering for updates on new projects",
-	ADDUSER_BUSY: "Adding user...",
-	GETAVAILABLEPLUGINBUNDLES_BUSY: "Loading available plugins, this can take a while...",
-	GETAVAILABLEPLUGINBUNDLES_DONE: "Done loading available plugins",
-	GETINSTALLEDPLUGINBUNDLES_BUSY: "Loading installed plugins, this can take a while...",
-	GETINSTALLEDPLUGINBUNDLES_DONE: "Done loading installed plugins",
-	INSTALLPLUGINBUNDLE_BUSY: "Installing plugin...",
-	INSTALLPLUGINBUNDLE_DONE: "Plugin successfully installed",
-	GETPLUGININFORMATION_BUSY: "Getting plugin information, this can take a while for large plugins...",
-	GETPLUGININFORMATION_DONE: "Plugin information successfully retrieved",
-	DOWNLOAD_BUSY: "Downloading model data...",
-	DOWNLOAD_DONE: "Model data downloaded",
-	LOGIN_DONE: "Login successful",
-	LOGOUT_DONE: "Logout successful",
-	UPDATEPROJECT_DONE: "Project successfully updated",
-	TRIGGERREVISIONSERVICE_BUSY: "Triggering service...",
-	TRIGGERREVISIONSERVICE_DONE: "Service triggered successfully",
-	INSTALLPLUGINBUNDLEFROMFILE_DONE: "Plugin bundle successfully installed from file",
-	CHECKINFROMURL_BUSY: "Checking in from URL...",
-	ERROR_REMOTE_METHOD_CALL: "Remote error (server probably down or not reachable)",
-	GETTING_MODEL_DATA: "Getting model data...",
-	MODEL_DATA_DONE: "Model data successfully downloaded...",
-	INITIATECHECKIN_BUSY: "Initiating checkin...",
-	GETTILECOUNTS_BUSY: "Counting tiles...",
-	LISTBOUNDINGBOXES_BUSY: "Loading bounding boxes...",
-	GETALLPROJECTS_BUSY: "Loading all projects..."
-};
